Rename functions
What?
Renamed all the functions starting with 'compile'

Why?
The functions are more or less behave like an interpreter,
rather than a compiler. As it doesn't parse entire message at one go.
Instead, it parses message token by token. If anywhere it finds invalid
token, it does not process rest of the message.